requirements.txt

amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.134.0
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
onnx==1.20.1
onnx-ir==0.2.0
onnx2torch==1.5.15
onnxruntime==1.24.2
onnxscript==0.6.2
opencv-python-headless==4.10.0.84
packaging==26.0
pandas==2.3.3
pillow==12.1.1
prompt_toolkit==3.0.52
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5

pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
redis==7.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
six==1.17.0
smmap==5.0.2
starlette==0.52.1

sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
torchvision==0.25.0

tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
captum==0.7.0
